Update 6-size.c file for 32-bit system

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	long long int b;
 	float f;
 
-	printf("Size of a char: %lu byte(s)", (unsigned long)sizeof(c));
-	printf("Size of an int: %lu byte(s)", (unsigned long)sizeof(i));
-	printf("Size of a long int: %lu byte(s)", (unsigned long)sizeof(a));
-	printf("Size of a long long int: %lu byte(s)", (unsigned long)sizeof(b));
-	printf("Size of a float: %lu byte(s)", (unsigned long)sizeof(f));
+	printf("Size of a char: %lu byte(s)", sizeof(c));
+	printf("Size of an int: %lu byte(s)", sizeof(i));
+	printf("Size of a long int: %lu byte(s)", sizeof(a));
+	printf("Size of a long long int: %lu byte(s)", sizeof(b));
+	printf("Size of a float: %lu byte(s)", sizeof(f));
 
 	return (0);
 }
